+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Important Changes:
 
+version 0.2.1:
+  - Astrometry solving in SimpleCalibPipeline
+  - Speedup FileManager
+
 version 0.2:
   - First version here: 0.2
   - Complete support for ROBO40 image processing pipeline